api/requirements.txt
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
pyyaml==5.3.1
uritemplate==3.0.1
django==3.1.4
dj-rest-auth[with_social]==2.1.2
django-allauth==0.44.0
model-bakery==1.2.1
responses==0.12.1
coverage==5.3.1
